Added user/progress/watched and user/progress/collected endpoints including optional title, sort order and extended parameters.
Added a SortType enumeration for the ordering of progress results.
Added a TvShowProgress entity.
Added basic tests for the new progress endpoints for "normal" and "full" information.

Added a type adapter for TvShowProgress.NextEpisode deserialization due to api using json primitive false when there is no episode, and a json object when there is.